pyproject.toml
[project]
name = "api-tester-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "sfelaco"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.23,<0.4.0)",
    "langgraph (>=0.3.30,<0.4.0)",
    "langchain-openai (>=0.3.13,<0.4.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
